Tightening Login Checks
These changes make sure that only accounts of a certain type can log in as that type and that the passwords are tied to the username we're checking (as opposed to the entire account table). Also, I've updated the DB to cascade delete for enduser when deleting a customer account.